Lint improvements (#824)
Adopts the following lints as a baseline (with a few exceptions):

    #![warn(
        clippy::mod_module_files,
        clippy::unwrap_used,
        missing_docs,
        rust_2018_idioms,
        unused_lifetimes,
        unused_qualifications
    )]

+++ pkcs5/src/pbes2/encryption.rs
@@ -171,7 +171,7 @@ impl EncryptionKey {
         pbkdf2::<Hmac<D>>(
             password,
             params.salt,
-            params.iteration_count as u32,
+            params.iteration_count,
             &mut buffer[..length],
         );
